Huge code fixes/refactoring
This commit refactors and reworks most of the `foohid` code:

* Fixes a few memory leaks happening in case of bad memory allocation,
(specifically inside the `foohid/foohid_userclient.cpp` file).
* It provides a better logging format.
* It allows creating virtual keyboard devices.
* It fixes an error that was not allowing to notify the caller
    in case of bad report descriptors

Please note that this is also not backward-compatible.
The CREATE, DESTROY and SEND user client methods do not require any
output buffer now, because the result of the operation is directly
returned to the caller.

Sorry for the non-atomic commit :)